fix: widen guest-controlled counts before +1 in COLUMNAR_V1 offset-size checks
outer_bytes/off_bytes/outer_offset_bytes computed (n + 1u) / (rows_to_dec
+ 1u) in uint32_t arithmetic before multiplying by sizeof(uint64_t). n
and rows_to_dec both derive from guest/frame-controlled counts (an inner
array/string element count decoded from offsets, or the frame header's
num_rows); a value of UINT32_MAX makes the +1 wrap to 0 in 32-bit
arithmetic before the multiply promotes it to 64-bit, so a malformed
frame could make the subsequent bounds check trivially pass and then
walk past the end of buf via unalignedLoad or the per-row loop.

Cast to uint64_t before the +1 at all three sites: the nested Array and
String branches in the COL_COMPLEX decode lambda, and the top-level
Array COL_COMPLEX branch.

+++ src/Formats/ColumnarV1Wire.h
@@ -631,7 +631,10 @@ inline MutableColumnPtr readColumnFromDesc(
     {
         if (const auto * arr_type = typeid_cast<const DataTypeArray *>(type.get()))
         {
-            uint64_t outer_bytes = (n + 1u) * sizeof(uint64_t);
+            // n comes from a guest-controlled row/element count; widen before the +1 or
+            // n == UINT32_MAX wraps the uint32_t addition to 0, making outer_bytes 0 and
+            // trivially passing the bounds check below.
+            uint64_t outer_bytes = (static_cast<uint64_t>(n) + 1u) * sizeof(uint64_t);
             if (p + outer_bytes > data_end)
                 throw Exception(ErrorCodes::INCORRECT_DATA,
                     "COLUMNAR_V1: COL_COMPLEX nested Array outer offsets out of bounds");
@@ -670,7 +673,8 @@ inline MutableColumnPtr readColumnFromDesc(
         }
         if (typeid_cast<const DataTypeString *>(type.get()))
         {
-            uint64_t off_bytes = (n + 1u) * sizeof(uint64_t);
+            // Same overflow hazard as the Array branch above: widen before the +1.
+            uint64_t off_bytes = (static_cast<uint64_t>(n) + 1u) * sizeof(uint64_t);
             if (p + off_bytes > data_end)
                 throw Exception(ErrorCodes::INCORRECT_DATA,
                     "COLUMNAR_V1: COL_COMPLEX String offsets out of bounds");
@@ -853,7 +857,9 @@ inline MutableColumnPtr readColumnFromDesc(
         {
             // WASM→CH sequential layout: outer uint64 offsets[rows+1] at data_offset,
             // followed immediately by nested writeComplexData-format data.
-            const uint64_t outer_offset_bytes = (rows_to_dec + 1u) * sizeof(uint64_t);
+            // rows_to_dec comes from the frame header's num_rows; widen before the +1 or
+            // rows_to_dec == UINT32_MAX wraps to 0, trivially passing the bounds check below.
+            const uint64_t outer_offset_bytes = (static_cast<uint64_t>(rows_to_dec) + 1u) * sizeof(uint64_t);
             if (outer_offset_bytes > desc.data_size)
                 throw Exception(ErrorCodes::INCORRECT_DATA,
                     "COLUMNAR_V1: COL_COMPLEX outer offsets exceed data_size: need={}, data_size={}",
